LIB: FormantCorrection: use float math internally (instead of doubles)

+++ lib/smformantcorrection.cc
@@ -15,7 +15,7 @@ FormantCorrection::FormantCorrection()
 }
 
 void
-FormantCorrection::set_ratio (double new_ratio)
+FormantCorrection::set_ratio (float new_ratio)
 {
   ratio = new_ratio;
 }
@@ -88,31 +88,31 @@ FormantCorrection::process_block (const AudioBlock& in_block, RTAudioBlock& out_
   auto emag = [&] (int i) {
     if (i > 0 && i < int (in_block.env.size()))
       return in_block.env_f (i);
-    return 0.0;
+    return 0.0f;
   };
-  auto emag_inter = [&] (double p) {
+  auto emag_inter = [&] (float p) {
     int ip = p;
-    double frac = p - ip;
+    float frac = p - ip;
     return emag (ip) * (1 - frac) + emag (ip + 1) * frac;
   };
-  auto set_mags = [&] (double *mags, size_t mags_count) {
+  auto set_mags = [&] (float *mags, size_t mags_count) {
     /* compute energy before formant correction */
-    double e1 = 0;
+    float e1 = 0;
     for (size_t i = 0; i < in_block.mags.size(); i++)
       {
-        double mag = in_block.mags_f (i);
+        float mag = in_block.mags_f (i);
         e1 += mag * mag;
       }
     /* compute energy after formant correction */
-    double e2 = 0;
+    float e2 = 0;
     for (size_t i = 0; i < mags_count; i++)
       {
-        double mag = mags[i];
+        float mag = mags[i];
         e2 += mag * mag;
       }
 
-    const double threshold = 1e-9;
-    double norm = (e2 > threshold) ? sqrt (e1 / e2) : 1;
+    const float threshold = 1e-9;
+    float norm = (e2 > threshold) ? sqrt (e1 / e2) : 1;
 
     /* generate normalized block mags */
     assert (out_block.freqs.size() == mags_count);
@@ -126,18 +126,18 @@ FormantCorrection::process_block (const AudioBlock& in_block, RTAudioBlock& out_
   if (mode == MODE_PRESERVE_SPECTRAL_ENVELOPE)
     {
       out_block.freqs.set_capacity (in_block.freqs.size());
-      const double e_tune_factor = 1 / in_block.env_f0;
-      double mags[in_block.freqs.size()];
+      const float e_tune_factor = 1 / in_block.env_f0;
+      float mags[in_block.freqs.size()];
       size_t count = 0;
 
       for (size_t i = 0; i < in_block.freqs.size(); i++)
         {
-          double freq = in_block.freqs_f (i) * e_tune_factor;
+          float freq = in_block.freqs_f (i) * e_tune_factor;
 
           if (freq > 0.5)
             {
-              double old_env_mag = emag_inter (freq);
-              double new_env_mag = emag_inter (freq * ratio);
+              float old_env_mag = emag_inter (freq);
+              float new_env_mag = emag_inter (freq * ratio);
 
               out_block.freqs.push_back (in_block.freqs[i]);
               mags[count++] = in_block.mags_f (i) / old_env_mag * new_env_mag;;
@@ -151,12 +151,12 @@ FormantCorrection::process_block (const AudioBlock& in_block, RTAudioBlock& out_
   else if (mode == MODE_HARMONIC_RESYNTHESIS)
     {
       int partials = std::min (sm_round_positive (max_partials / ratio) + 1, RESYNTH_MAX_PARTIALS);
-      double freqs[partials];
+      float freqs[partials];
       uint16_t ifreqs[partials];
-      double mags[partials];
+      float mags[partials];
       size_t mags_count = 0;
 
-      double ff = in_block.env_f0;
+      float ff = in_block.env_f0;
       if (fuzzy_frac > 1)
         {
           detune_factors.swap (next_detune_factors);
